Android/Logging: logs are now uploaded to Firebase
Needed to rework the Firebase component because the logging
task is not persistent. That means we need to query the existing
log entry when connecting. This should be verified how it behaves
when there is no internet connectivity (at that point, it should
be using the locally cached version, but need to ensure this is
handled properly and that isn't flushed within an hour).

Timings will also need to be adjusted to optimize sensor battery
life.