Android/Logging: migrate writing logs to DB into EmgImuManager
Since this is already the code that handles receiving and collating
the logs it should be in this code and keeps things simpler. Should
also make it work out the number of samples and when the first time
stamp will be and then prepare the log before downloading any samples.