Android/Service: better manage life cycle of log intents
There was a sneaky bug as the stopSelf(threadId) doesn't check that
all opened threadId's have been stopped, but only that you are stopping
the _most recent_. Thus we have to track them all as we want all to
complete before stopping the service and disconnecting from devices.

There might be a cleaner way to have a queue of outstanding tasks
to be completed in teh future, but since we just have one log
collection task this seems to work for now.