Android/Logging: sync timestamp and use sensor timestamps
This adds support for the extended RACP op codes to set the time
offset on the sensor. This will only be applied to the sensor once
per powerup. It also uses these timestamps from the sensor when
writing the log entries. It also changed how the logWritten event
occurs -- now it happens at the end of writing all the data to the
firebase logger but does not wait for an event indicating this is
done. The way it was previously done was causing crashes as if the
log ran into the next hour, it would sometimes shut down the service
before finishing the write.