Android: parent module includes all the libraries
This allows writing the separate games as libraries and including
them in a bundle. A build flavor might be used for excluding the
biofeedback components, or this could be disabled using Firebase
settings.

Ideally an instant-app architecture could be used, but Firebase does
not seem to play well with this, yet.