Android: repackage "core" as "config"
This better represents that this package is predominantly for
configuring the devices attached to the system (and will have
provisions for changing thresholds, etc), and that the
emgimuservice library is more core for that. All things will
be bundled in the APK.

Ultimately, there should still be a clean AIDL implemented between
the service and the activities to allow applications to be in
separate packages in the futures.