Android: move log fetching to JobDispatcher
Rather than running it in the main service (which seemed a bit elegant)
move this to the standalone JobDispatcher service. The rational for this
is that with Android O for background tasks you cannot wake up a service.
Triggering the Foreground service notification wasn't working realiably.

Still need to make sure that the log fetching doesn't wake up and then
close the connection to the device in normal operation.