Android: handle changing activities while connected
When an activity binds to the service and devices are
already connected, the BleMulticonnectProfileServiceReadyActivity
files an onDeviceConnected event when bound (see
ServiceConnection.onServiceConnected). Ideally at this level
the system would track if the device is also ready and
if so fire this event as that is much more pertinent for
how we interact with devices than if they are connected.

This tries to hack that functionality in without modifying the
core classes, but that requires the activity explicitly check
with a new method I wrote whenever they receive an onDeviceConnected
event.

It also uses the ugly approach of forwarding all queries
about a device to the service, which I'm keeping for now
but still think perhaps the EmgImuManager should be directly
exposed.