Android/StreamLogging: increase buffering and new thread
The separate thread may not be necessary compared to the
buffering, although the PipedInputStream documentation
specifies input and output streams should be handled in
separate threads (?unclear if UploadTask does this).
However, the larger buffer is the main thing that seems
to resolve things. Ultimately if this buffer fills things
will still block.

Of note using JSON is introducing a silly amount of data
bloat and is not appropriate as a long term data format.