Android/FlutterCow: never try and draw on invalid canvas
Crashlytics reports a crash for drawing to a null canvas. Most
likely it is when this wait loop gets interrupted and then
after logging the error attempts to draw anyway. Avoid this
behavior.

This I'm not certain _when_ this occurs there may be some other
crashes that arise after this.