Android/Log fetch: move back into the service
This essentially reverts 98b6647bfb2b42b3a46ca3dc7bb9d11b15203e57,
which should hopefully fix #72. This bug occurred when the device
was connected from the log fetch service and then the normal service
started in that time window. It created a window which drives the
sensor into a bad state and locks it up.