Android/Logging: startForeground issues
Force a notification whenever the log fetching task triggers
a startServiceForeground, because we must respond with
startForeground. I don't see an easy way to skip this, even
if we are already foregrounded or if BLE is disabled and we
don't even want to try.

Temporary workaround that seems to fix #77.