Android/Log: more work on RACP race conditions
Work towards fixing #79. I think the issue is various things
(such as trying to sync the device) which use the RACP characteristic
(probably inappropriately since it's mixing context) interrupt a
log fetch. The RACP has a certain level of granularity for commands
which is not enforced robustly on the android side.

This still doesn't fully resolve #79 as it is fairly easy to
trigger the condition where it doesn't stream.