Unity: check in PowerHammer project
Build process is still a little ugly. From this project
we then export the gradle project file to android which
then exports to the powerhammer subdirectory. The manifest
file needs to be recovered and then it can be compiled into
our architecture.

This supports the basic functionality to receive EMG messages
via a UnitySendMessage command on the Player.