Android/PowerHammer: fix ServiceConnection leak
This seems to arise from some mechanism that Unity binds a
service with intent Intent { act=com.google.android.gms.ads.identifier.service.START pkg=com.google.android.gms }
but then does not unbind it. My guess is this is because
we are not calling the onPause method on the player.