Android: support streaming attitude
Some of the pieces for streaming Accel/Gyro are in place
but they have not been implemented yet.

This is a lot of boilerplate code to relay messages through
numerous layers of indirection, which really indicates we
need a new model.

To subscript to data:
 ServiceBinder -> EmgImuManager (enable notifications)

Then receiving it
 EmgImuManager -> Service -> BroadcastReceiver -> Activity/ServiceHolder -> Callback

Really we need a model where we connect a listener directly and thta
process then enables the appropriate notification or disables it if
there are no remaining listeners. This is a fairly significant
refactoring, though.