Unity: add Attitude Tracking game stub
This takes the quaternion attitude and visualizes
it in Unity as a cube. This also extends the Unity
app to support an empty launcher scene from which
we can then select the sub-scene.

This is an alternative to having separate Unity
projects for each game, which would likely bloat
the app significantly. We could entirely split them
into separate games at some point.