Android/MVE: improve how we handle range and threshold
Separately store the min/max for range and the low and
high thresholds for clicking. This is also stored in a
per-device threshold now, which should be useful when
we have multiple devices.

This also fixes a quirk when changing thresholds did not
have any influence until the device reconnected.