Android: begin upgrade to Android-BLE-Library 2.0.4
This update has been long awaited. Significant work remains in
the EmgImuManager to become compliant with the new BleManager.
In addition, the BleMultiConnectProfile{Service, REadyActivity}
will also need to be updated with changes from the toolbox.

This might be a good idea to complete rework how activities
subscribe to messages from the service and improve the type
of abstraction we use.