Android/Config: respond to link loss
Also when device disconnected then do not remove from list.
However, the correct behavior depends on how the service is
managing the device list. If the service removes from the
list when we disconnect then we should actually copy that.

Really the problem is that the config page is meant to manage
the set of devices the service attempts to connect to, but
the code originated to reflect the set of devices we were
really bound to.