Android/Unity: callbacks now working
Still a bit hacky and more logic needs to move into
the service properly. Uses a timer to wait for sensor
to connect rather than callbacks since we aren't
exposing most of them.

Another hack is unity requires common-release.aar but
this required manually stripping the resources out of
the file to get it to link it. Still, overall progress
towards a cleaner barrier between the service and
consumers.