Android/Service: drop BleMulticonnectProfileService
That was a reasonable way to start but many of the design
patterns from it are no longer useful and in fact slow things
down. Most specifically, the handling of the device list. We
want this to be a persistent list of devices to attempt to
connect to. Ideally we tell the service what we want to hear
(e.g. subscribe to power) and it handles details like when the
device becomes ready and sends those commands.

The background log fetching is probably broken by these changes.