Android/UnityBridge: supports logging now
Note that logs provided this way are uploaded by the core service
so use the same username as that. It is probably best practice and
how the other apps should ultimately do it. Refactoring the
FirebaseGameLogger more cleanly after this would make sense.

Might also make sense to have a "finalize" method that sets the stop
time as opposed to updating it based on the most recent version, so
we know when apps crash. Similarly, performance field is not currently
logged.