Android: Updating Android code
Correcting bug in EmgImuService.java to compare
MAC Addresses of devices rather than strings of
the MACs. It seems like in newer versions, the
MAC is masked in strings as XX:XX:XX... etc so
the actual address needs to be compared. Also
committing other updates to gradle/manifest files