Android: Log timestamp sync upon log file open
There is already code in place to log the timestamp synchronization to
file when it happens in the `syncDevice()` function. This function is
called during initialization of the BLE connection, but it appears to
happen before the log file is created. Therefore, extra code needed to
be added to log the synchronized timestamp reference value to file as
soon as the file has been opened, so long as synchronization has
happened. This commit takes care of that.

[GUIS-41]