Android: Fix min/max EMG text display overlap
When the minimum EMG line exceeds the top of the screen, the min and
max lines are at the same location, so the text for each overlaps. I
repaired this and also added in a function to shift the minimum line
text to display above the minimum line when that line gets near the
bottom of the screen, similar to how we already do it with the max
text when the max line is near the top of the screen.

[GUIS-40]