Android: Removed minimum text logic
I had added in some logic to place the minimum line text above the
line instead of below when it got too close to the bottom of the
screen because it became unviewable. However, this caused the minimum
and maximum line text objects to overlap when both the minimum and
maximum were near the bottom of the screen. I removed this logic.

[GUIS-40]