Add minimal google-services.json files
These files contain the bare minimum required to allow the project to
build. Without these files, the build fails since these files are parsed
at that time. The entries that are in the file are dummy values.

This commit also includes a code change to catch an exception during
Firebase initialization. Since the entries in the JSON files are either
empty or dummy values, the initialization process fails. The exception
that gets thrown is caught, an error message gets printed to logcat, and
execution is allowed to continue.

[UTILS-74]